# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.CreateWODRequest:
    description: Request to create a new WOD
    properties:
      description:
        example: 15 min AMRAP of 10 KB swings, 15 box jumps
        type: string
      name:
        example: My Custom WOD
        type: string
      notes:
        type: string
      regime:
        example: 15 min
        type: string
      score_type:
        example: rounds+reps
        type: string
      source:
        example: custom
        type: string
      type:
        example: AMRAP
        type: string
      url:
        type: string
    type: object
  handler.ErrorResponse:
    description: Error response returned when a request fails
    properties:
      error:
        example: additional error details
        type: string
      message:
        example: Invalid credentials
        type: string
    type: object
...
  handler.MessageResponse:
    description: Generic message response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handler.UpdateWODRequest:
    description: Request to update an existing WOD
    properties:
      description:
        example: 21-15-9 thrusters and pull-ups
        type: string
      name:
        example: Updated WOD Name
        type: string
      notes:
        type: string
      regime:
        example: 21-15-9
        type: string
      score_type:
        example: time
        type: string
      source:
        example: custom
        type: string
      type:
        example: For Time
        type: string
      url:
        type: string
    type: object
...
  handler.WODListResponse:
    description: List of WODs
    properties:
      wods:
        items:
          $ref: '#/definitions/handler.WODResponse'
        type: array
    type: object
...
  handler.WODResponse:
    description: WOD details response
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        type: integer
      description:
        example: 21-15-9 thrusters and pull-ups
        type: string
      id:
        example: 1
        type: integer
      is_standard:
        example: true
        type: boolean
      name:
        example: Fran
        type: string
      notes:
        type: string
      regime:
        example: 21-15-9
        type: string
      score_type:
        example: time
        type: string
      source:
        example: CrossFit
        type: string
      type:
        example: For Time
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@actalog.com
    name: ActaLog Support
    url: https://github.com/johnzastrow/actalog/issues
  description: ActaLog is a mobile-first CrossFit workout tracker API. It provides
    endpoints for user authentication, workout logging, movement tracking, personal
    records, and administrative functions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://actalog.com/terms/
  title: ActaLog API
  version: 0.17.0
paths:
...
  /wods:
    get:
      consumes:
      - application/json
      description: Retrieve all WODs including standard and user's custom WODs
      parameters:
      - description: Max results (default 100)
        in: query
        name: limit
        type: integer
      - description: Skip N results (default 0)
        in: query
        name: offset
        type: integer
      - description: Return only standard WODs
        in: query
        name: standard
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of WODs
          schema:
            $ref: '#/definitions/handler.WODListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all WODs
      tags:
      - wods
    post:
      consumes:
      - application/json
      description: Create a new custom WOD (Workout of the Day)
      parameters:
      - description: WOD details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWODRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created WOD
          schema:
            $ref: '#/definitions/handler.WODResponse'
        "400":
          description: Invalid request body or missing name
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a custom WOD
      tags:
      - wods
  /wods/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a custom WOD (cannot delete standard WODs)
      parameters:
      - description: WOD ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: WOD deleted successfully
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid WOD ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden - cannot delete this WOD
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a WOD
      tags:
      - wods
    get:
      consumes:
      - application/json
      description: Retrieve a single WOD by its ID
      parameters:
      - description: WOD ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: WOD details
          schema:
            $ref: '#/definitions/handler.WODResponse'
        "400":
          description: Invalid WOD ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get WOD by ID
      tags:
      - wods
    put:
      consumes:
      - application/json
      description: Update an existing WOD (admin can update any, users only their
        own custom WODs)
      parameters:
      - description: WOD ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated WOD details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateWODRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated WOD
          schema:
            $ref: '#/definitions/handler.WODResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden - not your WOD
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a WOD
      tags:
      - wods
  /wods/mine:
    get:
      consumes:
      - application/json
      description: Retrieve only the authenticated user's custom WODs
      parameters:
      - description: Max results (default 100)
        in: query
        name: limit
        type: integer
      - description: Skip N results (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user's custom WODs
          schema:
            $ref: '#/definitions/handler.WODListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my custom WODs
      tags:
      - wods
  /wods/search:
    get:
      consumes:
      - application/json
      description: Search for WODs by name
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Matching WODs
          schema:
            $ref: '#/definitions/handler.WODListResponse'
        "400":
          description: Missing search query
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Search WODs
      tags:
      - wods
  /wods/standard:
    get:
      consumes:
      - application/json
      description: Retrieve only standard (built-in) WODs
      parameters:
      - description: Max results (default 100)
        in: query
        name: limit
        type: integer
      - description: Skip N results (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of standard WODs
          schema:
            $ref: '#/definitions/handler.WODListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List standard WODs
      tags:
      - wods
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"